+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 -   Added class override for `chevron` on `<InfoListItem>` and `<DrawerNavItem>`.
 -   Added `titleDivider` property onto `<DrawerNavGroup>`. ([#315](https://github.com/brightlayer-ui/react-component-library/issues/315))
 -   Added 1rem default padding to the root styles of `<EmptyState>`. ([#320](https://github.com/brightlayer-ui/react-component-library/issues/320))
+-   Added new property `unitSpace` to `<ChannelValue>` to manage spacing between the value and units. ([#350](https://github.com/brightlayer-ui/react-component-library/issues/350))
 
 ## v5.3.3 (December 6, 2021)